+++ CHANGELOG.md
@@ -33,6 +33,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   `BOOL()` function
 -   `JSON()` function
 -   `AS` keyword ([#24](https://github.com/codewithkyle/jsql/issues/24))
+-   nested object queries ([#12](https://github.com/codewithkyle/jsql/issues/12))
 
 ### Optimizations